Fix/remove lazy route imports (#827)
* use a single endpoint for getting settings and updating them, reducing http calls

* update settings and tests in a single call

* use `json_schema_extra` and fix defaults

* fix

* fix tests

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* add timeouts

* move imports to top of the file

* absolut import paths / isort

* nitpick comments

---------

+++ src/zndraw/settings.py
@@ -140,33 +140,33 @@ class PathTracing(SettingsBase):
 
     enabled: bool = Field(default=False, description="Enable GPU path tracing renderer")
 
-    min_samples: float = Field(
-        default=1.0,
-        ge=1.0,
+    min_samples: int = Field(
+        default=1,
+        ge=1,
         description="Minimum samples before displaying result",
         json_schema_extra={"format": "range"},
     )
 
-    samples: float = Field(
-        default=256.0,
-        ge=1.0,
-        le=10000.0,
+    samples: int = Field(
+        default=256,
+        ge=1,
+        le=10000,
         description="Maximum samples to render",
         json_schema_extra={"format": "range"},
     )
 
-    bounces: float = Field(
-        default=3.0,
-        ge=1.0,
-        le=32.0,
+    bounces: int = Field(
+        default=3,
+        ge=1,
+        le=32,
         description="Number of light bounces for global illumination",
         json_schema_extra={"format": "range"},
     )
 
-    tiles: float = Field(
-        default=1.0,
-        ge=1.0,
-        le=8.0,
+    tiles: int = Field(
+        default=1,
+        ge=1,
+        le=8,
         description="Rendering tile count (higher = less memory, slower)",
         json_schema_extra={"format": "range"},
     )
